requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
google-generativeai
pandas
wikipedia
python-dotenv
pydantic
